+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.11.0]
+
+### Added
+
+- Metadata profiles
+- Shapes in resource definitions
+- Metadata lables resolving
+
+### Fixed
+
+- Metadata with empty keywords
+- Pagination in index
+
+
 ## [1.10.0]
 
 ### Added
@@ -183,3 +198,4 @@ Initial version of client application for [FAIR Data Point] providing user inter
 [1.8.0]: /../../tree/v1.8.0
 [1.9.0]: /../../tree/v1.9.0
 [1.10.0]: /../../tree/v1.10.0
+[1.11.0]: /../../tree/v1.11.0